move get_notifications_page_partials_as_json and _get_notifications_dashboard_partials_data to dashboard.py
